refactors connection and auth modules
